+++ packages/playout/scheduler/program.py
@@ -4,6 +4,7 @@
 from database import Video
 from . import clock
 from vision.configuration import configuration
+from dateutil.parser import isoparse
 
 import datetime
 
@@ -13,17 +14,15 @@ class Program(object):
         datetime
     media_id
         integer
-    playback_offset
-        None or float in seconds
     playback_duration
         None or float in seconds
 
     TODO: perhaps specify a "pase" media_id or somethign?
     """
-    def __init__(self, media_id=None, program_start=None, playback_offset=None, title=None, playback_duration=None, data={}, filename=None, loop=False):
+
+    def __init__(self, media_id=None, program_start=None, title=None, playback_duration=None, data={}, filename=None, loop=False):
         self.program_start = program_start
         self.media_id = media_id
-        self.playback_offset = playback_offset
         self.playback_duration = playback_duration
         self.title = title
         self.data = data
@@ -68,17 +67,47 @@ class Program(object):
         return "<Program <Scheduled for [%s]-[%s]> <Video #%i: [%s]>>" % \
             (self.program_start, self.endTime(), self.media_id, self.title)
 
+class ScheduledVideo(Program):
+    def __init__(self, video, programStart: datetime.datetime, programEnd: datetime.datetime):
+        self.media_id = video.id
+        self.title = video.name
+        self.filename = None
+        self.loop = False
+        self.video = video
+        self.program_start = programStart
+        self.playback_duration = (programEnd - programStart).seconds
+
+    def asWeirdLegacyDict(self):
+        """ The schedule uses this dict format so just as an intermediate
+        step in cleaning up this code, this returns a dict with that format."""
+        duration = _millisecond_duration_from_endpoints(self.startTime, self.endTime)
+        video = {
+                'broadcast_location': self.video.id,
+                'duration': duration,
+                'name': self.video.name,
+                'starttime': self.startTime,
+                }
+        return video
+
     @classmethod
     def fromWeirdLegacyDict(cls, weirdDict, playbackOffset = 0.0, additionalData = None):
         """ Loads the weird dict generated by ScheduledVideo """
         d = weirdDict
         return cls(
-                    media_id=d["broadcast_location"],
-                    program_start=d["starttime"],
-                    playback_offset=playbackOffset,
-                    playback_duration=d["duration"] / 1000.,
-                    title=d["name"],
-                    # unused:
-                    # endtime, video_id, header, schedule_reagion....
-                    data=additionalData
-                    )
+            media_id=d["broadcast_location"],
+            program_start=d["starttime"],
+            playback_duration=d["duration"] / 1000.,
+            title=d["name"],
+            # unused:
+            # endtime, video_id, header, schedule_reagion....
+            data=additionalData
+        )
+
+    @classmethod
+    def fromGraphNode(cls, node):
+        entry = node['node']
+        startTime = isoparse(entry['starttime']).astimezone(tz=None).replace(tzinfo=None)
+        endTime = isoparse(entry['endtime']).astimezone(tz=None).replace(tzinfo=None)
+        video = Video(entry['videoId'])
+        video.name = entry['videoName']
+        return cls(video, startTime, endTime)
